FreelancerM/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-widget-tweaks==1.5
channels==4.0.0
channels-redis==4.2.0
django-allauth==0.63.0
requests==2.32.5
yarg==0.1.10
docopt==0.6.2
pipreqs==0.4.13
dj-database-url==2.2.0
gunicorn==22.0.0
celery==5.3.6
redis==5.0.1

eventlet==0.33.3
